refacto: moved chip variant from core to uikit after PR review
