+++ Sources/VitaminUIKit/Components/Chip/VitaminChipVariant.swift
